backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
numpy
python-multipart
